+++ apps/api/src/app/access/access.controller.ts
@@ -87,7 +87,7 @@ export class AccessController {
     }
 
     try {
-      return this.accessService.createAccess({
+      return await this.accessService.createAccess({
         alias: data.alias || undefined,
         granteeUser: data.granteeUserId
           ? { connect: { id: data.granteeUserId } }
@@ -155,7 +155,7 @@ export class AccessController {
     }
 
     try {
-      return this.accessService.updateAccess({
+      return await this.accessService.updateAccess({
         data: {
           alias: data.alias,
           granteeUser: data.granteeUserId