+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Refactored the deprecated animation providers (`provideAnimations()` and `provideNoopAnimations()`)
 - Upgraded `chartjs-chart-treemap` from version `3.1.0` to `4.2.0`
 
 ### Fixed