+++ CHANGELOG.md
@@ -169,6 +169,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improved the language localization for Chinese (`zh`)
 - Improved the language localization for German (`de`)
 
+### Fixed
+
+- Fixed an issue in the portfolio details endpoint that incorrectly reported an error when a holding contained undefined optional values
+
 ## 3.22.0 - 2026-07-08
 
 ### Added